Github action, code improvements: - npm install + test github action - code cleanup - program should not exit when an eth transaction fails in a cron sequence
