+++ .github/workflows/build.yml
@@ -1,0 +1,24 @@
+name: Build
+
+on: [push]
+
+jobs:
+    build:
+        runs-on: ubuntu-latest
+
+        strategy:
+            matrix:
+                node-version: [14.x, 15.x, 16.x]
+
+        steps:
+            - uses: actions/checkout@v1
+            - name: Use Node.js ${{ matrix.node-version }}
+              uses: actions/setup-node@v1
+              with:
+                  node-version: ${{ matrix.node-version }}
+            - name: run npm install and test
+              run: |
+                  npm ci
+                  npm test -- --coverage
+              env:
+                  CI: true
