+++ package.json
@@ -7,8 +7,8 @@
     "test": "jest",
     "start": "node dist/src/index.js",
     "build": "tsc -p .",
-    "dev": "nodemon --exec ts-node src/index.ts",
-    "dev:single": "nodemon --exec ts-node src/index.ts singleRun"
+    "dev": "NODE_ENV=development nodemon --exec ts-node src/index.ts",
+    "dev:single": "NODE_ENV=development nodemon --exec ts-node src/index.ts singleRun"
   },
   "author": "",
   "license": "ISC",
